split synonym maps client
