+++ sdk/search/azure-search-documents/samples/async_samples/sample_synonym_map_operations_async.py
@@ -27,11 +27,11 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 from azure.core.credentials import AzureKeyCredential
 from azure.search.documents.aio import SearchServiceClient
 
-service_client = SearchServiceClient(service_endpoint, AzureKeyCredential(key))
+client = SearchServiceClient(service_endpoint, AzureKeyCredential(key)).get_synonym_maps_client()
 
 async def create_synonym_map():
     # [START create_synonym_map_async]
-    result = await service_client.create_synonym_map("test-syn-map", [
+    result = await client.create_synonym_map("test-syn-map", [
         "USA, United States, United States of America",
         "Washington, Wash. => WA",
     ])
@@ -40,14 +40,14 @@ async def create_synonym_map():
 
 async def get_synonym_maps():
     # [START get_synonym_maps_async]
-    result = await service_client.get_synonym_maps()
+    result = await client.get_synonym_maps()
     names = [x["name"] for x in result]
     print("Found {} Synonym Maps in the service: {}".format(len(result), ", ".join(names)))
     # [END get_synonym_maps_async]
 
 async def get_synonym_map():
     # [START get_synonym_map_async]
-    result = await service_client.get_synonym_map("test-syn-map")
+    result = await client.get_synonym_map("test-syn-map")
     print("Retrived Synonym Map 'test-syn-map' with synonyms")
     for syn in result["synonyms"]:
         print("    {}".format(syn))
@@ -55,7 +55,7 @@ async def get_synonym_map():
 
 async def delete_synonym_map():
     # [START delete_synonym_map_async]
-    await service_client.delete_synonym_map("test-syn-map")
+    await client.delete_synonym_map("test-syn-map")
     print("Synonym Map 'test-syn-map' deleted")
     # [END delete_synonym_map_async]
 
@@ -64,7 +64,7 @@ async def main():
     await get_synonym_maps()
     await get_synonym_map()
     await delete_synonym_map()
-    await service_client.close()
+    await client.close()
 
 if __name__ == '__main__':
     loop = asyncio.get_event_loop()
