+++ sdk/search/azure-search-documents/samples/sample_synonym_map_operations.py
@@ -26,11 +26,11 @@ key = os.getenv("AZURE_SEARCH_API_KEY")
 from azure.core.credentials import AzureKeyCredential
 from azure.search.documents import SearchServiceClient
 
-service_client = SearchServiceClient(service_endpoint, AzureKeyCredential(key))
+client = SearchServiceClient(service_endpoint, AzureKeyCredential(key)).get_synonym_maps_client()
 
 def create_synonym_map():
     # [START create_synonym_map]
-    result = service_client.create_synonym_map("test-syn-map", [
+    result = client.create_synonym_map("test-syn-map", [
         "USA, United States, United States of America",
         "Washington, Wash. => WA",
     ])
@@ -39,14 +39,14 @@ def create_synonym_map():
 
 def get_synonym_maps():
     # [START get_synonym_maps]
-    result = service_client.get_synonym_maps()
+    result = client.get_synonym_maps()
     names = [x["name"] for x in result]
     print("Found {} Synonym Maps in the service: {}".format(len(result), ", ".join(names)))
     # [END get_synonym_maps]
 
 def get_synonym_map():
     # [START get_synonym_map]
-    result = service_client.get_synonym_map("test-syn-map")
+    result = client.get_synonym_map("test-syn-map")
     print("Retrived Synonym Map 'test-syn-map' with synonyms")
     for syn in result["synonyms"]:
         print("    {}".format(syn))
@@ -54,7 +54,7 @@ def get_synonym_map():
 
 def delete_synonym_map():
     # [START delete_synonym_map]
-    service_client.delete_synonym_map("test-syn-map")
+    client.delete_synonym_map("test-syn-map")
     print("Synonym Map 'test-syn-map' deleted")
     # [END delete_synonym_map]
